+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Do not stop follow importer task on validation error.
+
 ## [2.18.0] - 2024-05-07
 
 ### Added